requirements.txt
streamlit
CoolProp
numpy
Pillow